+++ Projects/Skylicht/Engine/Source/Material/Shader/CShaderManager.cpp
@@ -45,7 +45,7 @@ namespace Skylicht
 
 	void CShaderManager::releaseAll()
 	{
-		for (u32 i = 0, n = m_listShader.size(); i < n; i++)
+		for (u32 i = 0, n = (u32)m_listShader.size(); i < n; i++)
 			m_listShader[i]->drop();
 
 		m_listShader.clear();
@@ -239,7 +239,7 @@ namespace Skylicht
 
 	CShader* CShaderManager::getShaderByName(const char *name)
 	{
-		for (u32 i = 0, n = m_listShader.size(); i < n; i++)
+		for (u32 i = 0, n = (u32)m_listShader.size(); i < n; i++)
 		{
 			if (m_listShader[i]->getName() == name)
 				return m_listShader[i];
@@ -250,7 +250,7 @@ namespace Skylicht
 
 	CShader* CShaderManager::getShaderByPath(const char *path)
 	{
-		for (u32 i = 0, n = m_listShader.size(); i < n; i++)
+		for (u32 i = 0, n = (u32)m_listShader.size(); i < n; i++)
 		{
 			if (m_listShader[i]->getShaderPath() == path)
 				return m_listShader[i];
@@ -261,7 +261,7 @@ namespace Skylicht
 
 	CShader* CShaderManager::getShaderByID(int id)
 	{
-		for (u32 i = 0, n = m_listShader.size(); i < n; i++)
+		for (u32 i = 0, n = (u32)m_listShader.size(); i < n; i++)
 		{
 			if (m_listShader[i]->getMaterialRenderID() == id)
 				return m_listShader[i];
